requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
firebase-admin==6.1.0
google-cloud-firestore==2.11.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
pydantic==2.4.2
pydantic[email]==2.4.2
pydantic-settings==2.0.3
python-dateutil==2.8.2